moves base terms to their own file
