+++ src/parser/mod.rs
@@ -1,3 +1,4 @@
+pub mod base_terms;
 pub mod parser;
 
 // TODO: trim this down
